# InplaceStringBuilder: better performance