--- a/src/Microsoft.Extensions.Primitives/Resources.resx
+++ b/src/Microsoft.Extensions.Primitives/Resources.resx
@@ -120,4 +120,7 @@
   <data name="Argument_InvalidOffsetLength" xml:space="preserve">
     <value>Offset and length are out of bounds for the string or length is greater than the number of characters from index to the end of the string.</value>
   </data>
+  <data name="Capacity_CannotChangeAfterWriteStarted" xml:space="preserve">
+    <value>Cannot change capacity after write started.</value>
+  </data>
 </root>