--- a/src/Microsoft.Extensions.Primitives/ThrowHelper.cs
+++ b/src/Microsoft.Extensions.Primitives/ThrowHelper.cs
@@ -23,6 +23,11 @@ namespace Microsoft.Extensions.Primitives
             throw new ArgumentException(GetResourceText(resource));
         }
 
+        internal static void ThrowInvalidOperationException(ExceptionResource resource)
+        {
+            throw new InvalidOperationException(GetResourceText(resource));
+        }
+
         internal static ArgumentNullException GetArgumentNullException(ExceptionArgument argument)
         {
             return new ArgumentNullException(GetArgumentName(argument));
@@ -68,11 +73,14 @@ namespace Microsoft.Extensions.Primitives
         text,
         start,
         count,
-        index
+        index,
+        value,
+        capacity
     }
 
     internal enum ExceptionResource
     {
-        Argument_InvalidOffsetLength
+        Argument_InvalidOffsetLength,
+        Capacity_CannotChangeAfterWriteStarted
     }
 }