# Description

* covered current behavior with tests + added some test to cover the public api
* instead of throwing exceptions directly, they are thrown via _ThrowHelper_ or private methods, thus making the code for the method smaller
* decreased the size of the struct, by removing two fields (`_capacity` and `_writing`)
* `_value` gets initialized in the ctor, potentially reinitialized in setting the _Capacity_
* made _Append_-methods to inline

A variant with `char[]` instead of `string` as `_value` was tried, but the results didn't satisfy.

# Benchmarks

## Append String, StringSegment and Char

[Benchmark-Code](https://github.com/gfoidl/Benchmarks/blob/a954945e44ec2b936076de51106b5968e9fecfd9/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/Benchmarks/AppendRealisticBenchmarks.cs)

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.125)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742191 Hz, Resolution=364.6719 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|                                 Method |      Mean |     Error |    StdDev | Scaled | ScaledSD |
|--------------------------------------- |----------:|----------:|----------:|-------:|---------:|
|                                Default |  74.63 ns | 1.5745 ns | 2.5426 ns |   1.00 |     0.00 |
| Default_ThrowHelper_AggressiveInlining |  42.83 ns | 0.5936 ns | 0.5262 ns |   0.57 |     0.02 |
|                    Default_ThrowHelper |  54.69 ns | 0.3238 ns | 0.2704 ns |   0.73 |     0.02 |
|                              CharArray | 178.26 ns | 1.7530 ns | 1.6397 ns |   2.39 |     0.08 |


## Append String

[Benchmark-Code](https://github.com/gfoidl/Benchmarks/blob/a954945e44ec2b936076de51106b5968e9fecfd9/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/Benchmarks/AppendStringBenchmarks.cs)

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.125)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742191 Hz, Resolution=364.6719 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|                                 Method |       Mean |     Error |    StdDev | Scaled | ScaledSD |
|--------------------------------------- |-----------:|----------:|----------:|-------:|---------:|
|                                Default |   453.2 ns |  8.977 ns | 14.749 ns |   1.00 |     0.00 |
| Default_ThrowHelper_AggressiveInlining |   278.1 ns |  4.503 ns |  4.212 ns |   0.61 |     0.02 |
|                    Default_ThrowHelper |   340.9 ns |  4.625 ns |  4.100 ns |   0.75 |     0.03 |
|                              CharArray | 1,451.3 ns | 19.951 ns | 16.660 ns |   3.21 |     0.11 |


## Append Char

[Benchmark-Code](https://github.com/gfoidl/Benchmarks/blob/a954945e44ec2b936076de51106b5968e9fecfd9/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/Benchmarks/AppendCharBenchmarks.cs)

``` ini

BenchmarkDotNet=v0.10.12, OS=Windows 10 Redstone 3 [1709, Fall Creators Update] (10.0.16299.125)
Intel Core i7-7700HQ CPU 2.80GHz (Kaby Lake), 1 CPU, 8 logical cores and 4 physical cores
Frequency=2742191 Hz, Resolution=364.6719 ns, Timer=TSC
.NET Core SDK=2.1.4
  [Host]     : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT
  DefaultJob : .NET Core 2.0.5 (Framework 4.6.26020.03), 64bit RyuJIT


```
|              Method |     Mean |     Error |    StdDev | Scaled | ScaledSD |
|-------------------- |---------:|----------:|----------:|-------:|---------:|
|             Default | 553.2 ns | 10.889 ns | 18.193 ns |   1.00 |     0.00 |
| Default_ThrowHelper | 242.9 ns |  3.939 ns |  3.684 ns |   0.44 |     0.02 |
|           CharArray | 213.0 ns |  4.288 ns |  8.158 ns |   0.39 |     0.02 |


# Discussion

## AggressiveInlining

Perf with _aggressive inling_ is a lot better than without. So I'll kept it, although the caller might get bigger in code size.

## Initialization of _value

`_value` gets initialized in the ctor, when a capacity is given. If a new capacity is set via the property, so a new string is initialized, throwing away a potential previous initialized string (_value). This simplifies the checks while appending.

The current version initialized the `_value` on its first use, so there is no throw-away even if the capacity is set by the property.

I think the "new bahavior" is it worth, though setting the capacity via ctor and later via property may be a less common approach.
Or in other words: when the capacity is set via ctor xor set via property, there is no throw-away either.