@rynowak 2018-02-21
@pakrym - are you most familiar with this type? want to take a look at this **SPICY** pr?