@pakrym 2018-02-21
I like this, my only concern is the following pattern:

```
var sb = new InplaceStringBuilder(5);
if (...)
sb.Capacity += 7;
if (...)
sb.Capacity += 3;
```

I looked through asp.net core code and it seems we don't use it but if anyone did things would get much slower.

How expensive is the `if ()` check in append?