@gfoidl 2018-02-21
> How expensive is the `if ()` check in append?

This depends on the usage pattern. Depending on that the JIT will or will not inline some methods and the property setter (for Capacity), resulting in +5% to -20% compared to the code of this PR.

### Append String

|              Method |     Mean |    Error |   StdDev | Scaled | ScaledSD |
|-------------------- |---------:|---------:|---------:|-------:|---------:|
|             [Default](https://github.com/gfoidl/Benchmarks/blob/39a3ea8621a6d0885a7ca6c8981a10e5583b34b8/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/InplaceStringBuilders/InplaceStringBuilder0.cs) | 425.6 ns | 3.895 ns | 3.041 ns |   1.00 |     0.00 |
|             [This_PR](https://github.com/gfoidl/Benchmarks/blob/39a3ea8621a6d0885a7ca6c8981a10e5583b34b8/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/InplaceStringBuilders/InplaceStringBuilder00.cs) | 282.4 ns | 5.741 ns | 5.370 ns |   0.66 |     0.01 |
| [This_PR_wo_pre_init](https://github.com/gfoidl/Benchmarks/blob/39a3ea8621a6d0885a7ca6c8981a10e5583b34b8/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/InplaceStringBuilders/InplaceStringBuilder02.cs) | 291.5 ns | 5.920 ns | 4.944 ns |   0.69 |     0.01 |
|            [New_Idea](https://github.com/gfoidl/Benchmarks/blob/39a3ea8621a6d0885a7ca6c8981a10e5583b34b8/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/InplaceStringBuilders/InplaceStringBuilder01.cs) | 264.2 ns | 5.662 ns | 6.520 ns |   0.62 |     0.02 |

### Append String, StringSegment and Char

|              Method |     Mean |     Error |    StdDev | Scaled |  Gen 0 | Allocated |
|-------------------- |---------:|----------:|----------:|-------:|-------:|----------:|
|             [Default](https://github.com/gfoidl/Benchmarks/blob/39a3ea8621a6d0885a7ca6c8981a10e5583b34b8/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/InplaceStringBuilders/InplaceStringBuilder0.cs) | 68.31 ns | 0.7611 ns | 0.6747 ns |   1.00 | 0.0151 |      48 B |
|             [This_PR](https://github.com/gfoidl/Benchmarks/blob/39a3ea8621a6d0885a7ca6c8981a10e5583b34b8/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/InplaceStringBuilders/InplaceStringBuilder00.cs) | 44.31 ns | 0.3620 ns | 0.3023 ns |   0.65 | 0.0152 |      48 B |
| [This_PR_wo_pre_init](https://github.com/gfoidl/Benchmarks/blob/39a3ea8621a6d0885a7ca6c8981a10e5583b34b8/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/InplaceStringBuilders/InplaceStringBuilder02.cs) | 34.11 ns | 0.4492 ns | 0.4202 ns |   0.50 | 0.0152 |      48 B |
|            [New_Idea](https://github.com/gfoidl/Benchmarks/blob/39a3ea8621a6d0885a7ca6c8981a10e5583b34b8/aspnet/Common/InplaceStringBuilder/InplaceStringBuilder/InplaceStringBuilders/InplaceStringBuilder01.cs) | 34.92 ns | 0.2254 ns | 0.1998 ns |   0.51 | 0.0152 |      48 B |

`New_Idea` is this PR + tracking of the capacity as field, thus allowing the JIT to inline the access to `Capacity`. 

The differences in perf results from the JIT's inlining behavior.

My favorite is now `New_Idea`, but this has the same concern as https://github.com/aspnet/Common/pull/313#issuecomment-367395930
As stated in the description of the PR, I don't think this does matter because the throw-away of the pre-init string won't produce a lot garbage -- under the assumption that the capacity isn't set frequently and to big values.

The variant without the pre-initialization is fine, too. Just depending on the usage-pattern it may be a little slower than a pre-init variant.

So what to do?