@pakrym 2018-02-21
Let's go with lazy init option, I don't want to throw people that increment `Capacity` off the cliff.